moving solely to stream processor (will have processing rates handled within, with stream output acting as time-normalizing middleware)
